requirements.txt
a2a-sdk==0.3.19
agentbeats==1.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
archspec
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
backports.tarfile==1.2.0
banks==2.2.0
bcrypt==5.0.0
beartype==0.22.7
beautifulsoup4==4.14.3
boltons
Brotli
bs4==0.0.2
cachetools==6.2.2
certifi
cffi==2.0.0
charset-normalizer
click==8.3.1
colorama==0.4.6
cryptography
curl_cffi==0.13.0
cyclopts==4.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distro
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
earthshaker
email-validator==2.3.0
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.123.0
fastmcp==2.13.1
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
frozendict
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.15.0
grpcio==1.67.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna
importlib_metadata==8.7.0
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
litellm==1.80.7
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-llama-cpp==0.5.1
llama-index-llms-openai==0.6.10
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-faiss==0.5.1
llama-index-workflows==2.11.5
llama-parse==0.6.54
llama_cpp_python==0.3.16
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.22.0
mdurl==0.1.2
menuinst
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
openai==2.8.1
openai-agents==0.6.1
openapi-pydantic==0.5.1
packaging
pandas==2.2.3
paramiko==4.0.0
pathable==0.4.4
pathvalidate==3.3.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
pluggy
postgrest==2.24.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat
pycparser
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pypdf==6.4.0
pyperclip==1.11.0
PySocks
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml
ruamel.yaml.clib
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
SecretStorage==3.5.0
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
transformers==4.57.3
truststore
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zstandard
